+++ CHANGELOG.md
@@ -2,7 +2,32 @@
 
 This file is used to list changes made in each version of the gem.
 
-# 0.2.1
+# 0.4.0
+
+Fixes, updates and health_check release.
+
+- Update generators for Docker and Gitlab CI.
+- Move all generators under the `modulorails` namespace.
+- Add dependency to `health_check` gem.
+- Fix error on database configuration validator when no database.yml exists.
+- Rescue if httparty can't post to configuration.endpoint.
+- Add dockerfiles to test on many Ruby versions.
+- Add appraisal to test on many Rails versions.
+
+# 0.3.0
+
+Docker release.
+
+- Add generator for Docker.
+- Use templates for Gitlabci generator. 
+
+# 0.2.3
+
+Gitlab-ci generator.
+
+- Fixes the Ruby version put in the generated `.gitlab-ci.yml`.
+
+# 0.2.2
 
 Auto-update fixes.